installer: split plan.rs into plan, run, say
